Moved logger logic from index.ts to application.ts
