+++ src/index.ts
@@ -1,54 +1,12 @@
 import {ApplicationConfig, MyriadApiApplication} from './application';
 import {config} from './config';
-import {LoggingBindings, WinstonLoggerOptions, 
-  WINSTON_FORMAT, WINSTON_TRANSPORT,
-  WinstonFormat,
-  WinstonTransports
-} from '@loopback/logging';
-import {format} from 'winston';
-import {extensionFor} from '@loopback/core';
 
 export * from './application';
 
 export async function main(options: ApplicationConfig = {}) {
   const app = new MyriadApiApplication(options);
-  app.configure(LoggingBindings.COMPONENT).to({
-    enableFluent: false,
-    enableHttpAccessLog: true,
-  });
-
-  app.configure<WinstonLoggerOptions>(LoggingBindings.WINSTON_LOGGER).to({
-    level: 'info',
-    format: format.json(),
-    defaultMeta: {framework: 'LoopBack'},
-  });
-
-  const myFormat: WinstonFormat = format((info, opts) => {
-    console.log(info);
-    return false;
-  })();
-  
-  app
-    .bind('logging.winston.formats.myFormat')
-    .to(myFormat)
-    .apply(extensionFor(WINSTON_FORMAT));
-    app
-    .bind('logging.winston.formats.colorize')
-    .to(format.colorize())
-    .apply(extensionFor(WINSTON_FORMAT));
-  
-  const consoleTransport = new WinstonTransports.Console({
-    level: 'info',
-    format: format.combine(format.colorize(), format.simple()),
-  });
-  app
-    .bind('logging.winston.transports.console')
-    .to(consoleTransport)
-    .apply(extensionFor(WINSTON_TRANSPORT));
   await app.boot();
   await app.start();
-
-  
 
   const url = app.restServer.url;
   console.log(`Server is running at ${url}`);
